feat(Calculator): 덧셈 기능 구현 및 long 타입 지원

+++ src/main/java/calculator/presentation/ApplicationRunner.java
@@ -16,7 +16,7 @@ public class ApplicationRunner {
 
     public void run() {
         String inputString = input.readExpression();
-        int result = calculator.calculate(inputString);
+        long result = calculator.calculate(inputString);
         output.printResult(result);
     }
